requirements.txt
Django==3.2.6
django_filter==21.1
django_phonenumber_field==6.0.0
django_templated_mail==1.1.1
djangorestframework==3.14.0
djoser==2.1.0
drf_spectacular==0.22.0
python-dotenv==1.0.0
templated_mail==0.4.0.2
